+++ mb/core/BufferAttribute.hpp
@@ -35,10 +35,7 @@ namespace MB
 		MB_API
 		BufferAttribute(std::vector<float> arr, unsigned int size);
 		MB_API
-		void array(const std::vector<float> arr)
-		{
-			this->_arr = arr;
-		}
+		void array(const std::vector<float> arr);
 		MB_API
         unsigned int size() const;
 		MB_API