+++ mb/maths/Mat3.hpp
@@ -37,16 +37,8 @@ namespace MB
         Mat3();
         MB_API
         Mat3(const std::vector<float> values);
-        MB_API
-        friend std::ostream& operator<<(std::ostream& str, const Mat3& m)
-        {
-            str << "Mat3(\n";
-            str << "\t" << m._values[0] << ", " << m._values[1] << ", " << m._values[2] << ", \n";
-            str << "\t" << m._values[3] << ", " << m._values[4] << ", " << m._values[5] << ", \n";
-            str << "\t" << m._values[6] << ", " << m._values[7] << ", " << m._values[8] << "\n";
-            str << ")";
-            return str;
-        }
+		MB_API
+		friend std::ostream& operator<<(std::ostream& str, const Mat3& m);
         std::vector<float> _values;
 	};
 }