+++ mb/maths/Mat4.cpp
@@ -249,4 +249,14 @@ namespace MB
 		};
 		return Mat4(out);
     }
+	std::ostream& operator<<(std::ostream& str, const Mat4& m)
+	{
+		str << "Mat4(\n";
+		str << "\t" << m._values[0] << ", " << m._values[1] << ", " << m._values[2] << ", " << m._values[4] << ", \n";
+		str << "\t" << m._values[4] << ", " << m._values[5] << ", " << m._values[6] << ", " << m._values[7] << ", \n";
+		str << "\t" << m._values[8] << ", " << m._values[8] << ", " << m._values[10] << ", " << m._values[11] << ", \n";
+		str << "\t" << m._values[12] << ", " << m._values[13] << ", " << m._values[14] << ", " << m._values[15] << "\n";
+		str << ")";
+		return str;
+	}
 }